+++ driver-core/src/main/java/com/datastax/driver/core/Row.java
@@ -565,7 +565,7 @@ public class Row {
      * @return the value of the {@code i}th column in this row as a list of
      * {@code elementsClass} objects. If the value is NULL, an empty list is
      * returned (note that Cassandra makes no difference between an empty list
-     * and column of type list that is not set).
+     * and column of type list that is not set). The returned list is immutable.
      *
      * @throws IndexOutOfBoundsException if {@code i < 0 || i >= this.columns().size()}.
      * @throws InvalidTypeException if column {@code i} is not a list or if its
@@ -585,7 +585,7 @@ public class Row {
         if (value == null)
             return Collections.<T>emptyList();
 
-        return (List<T>)type.codec().deserialize(value);
+        return Collections.unmodifiableList((List<T>)type.codec().deserialize(value));
     }
 
     /**
@@ -596,7 +596,7 @@ public class Row {
      * @return the value of the {@code i}th column in this row as a list of
      * {@code elementsClass} objects. If the value is NULL, an empty list is
      * returned (note that Cassandra makes no difference between an empty list
-     * and column of type list that is not set).
+     * and column of type list that is not set). The returned list is immutable.
      *
      * @throws IllegalArgumentException if {@code name} is not part of the
      * ResultSet this row is part of, i.e. if {@code !this.columns().names().contains(name)}.
@@ -615,7 +615,7 @@ public class Row {
      * @return the value of the {@code i}th column in this row as a set of
      * {@code elementsClass} objects. If the value is NULL, an empty set is
      * returned (note that Cassandra makes no difference between an empty set
-     * and column of type set that is not set).
+     * and column of type set that is not set). The returned set is immutable.
      *
      * @throws IndexOutOfBoundsException if {@code i < 0 || i >= this.columns().size()}.
      * @throws InvalidTypeException if column {@code i} is not a set or if its
@@ -635,7 +635,7 @@ public class Row {
         if (value == null)
             return Collections.<T>emptySet();
 
-        return (Set<T>)type.codec().deserialize(value);
+        return Collections.unmodifiableSet((Set<T>)type.codec().deserialize(value));
     }
 
     /**
@@ -646,7 +646,7 @@ public class Row {
      * @return the value of the {@code i}th column in this row as a set of
      * {@code elementsClass} objects. If the value is NULL, an empty set is
      * returned (note that Cassandra makes no difference between an empty set
-     * and column of type set that is not set).
+     * and column of type set that is not set). The returned set is immutable.
      *
      * @throws IllegalArgumentException if {@code name} is not part of the
      * ResultSet this row is part of, i.e. if {@code !this.columns().names().contains(name)}.
@@ -666,7 +666,8 @@ public class Row {
      * @return the value of the {@code i}th column in this row as a map of
      * {@code keysClass} to {@code valuesClass} objects. If the value is NULL,
      * an empty map is returned (note that Cassandra makes no difference
-     * between an empty map and column of type map that is not set).
+     * between an empty map and column of type map that is not set). The
+     * returned map is immutable.
      *
      * @throws IndexOutOfBoundsException if {@code i < 0 || i >= this.columns().size()}.
      * @throws InvalidTypeException if column {@code i} is not a map, if its
@@ -688,7 +689,7 @@ public class Row {
         if (value == null)
             return Collections.<K, V>emptyMap();
 
-        return (Map<K, V>)type.codec().deserialize(value);
+        return Collections.unmodifiableMap((Map<K, V>)type.codec().deserialize(value));
     }
 
     /**
@@ -700,7 +701,8 @@ public class Row {
      * @return the value of the {@code i}th column in this row as a map of
      * {@code keysClass} to {@code valuesClass} objects. If the value is NULL,
      * an empty map is returned (note that Cassandra makes no difference
-     * between an empty map and column of type map that is not set).
+     * between an empty map and column of type map that is not set). The
+     * returned map is immutable.
      *
      * @throws IllegalArgumentException if {@code name} is not part of the
      * ResultSet this row is part of, i.e. if {@code !this.columns().names().contains(name)}.