+++ driver-core/src/main/java/com/datastax/driver/core/ConsistencyLevel.java
@@ -19,16 +19,17 @@ import com.datastax.driver.core.exceptions.DriverInternalError;
 
 public enum ConsistencyLevel {
 
-    ANY         (0),
-    ONE         (1),
-    TWO         (2),
-    THREE       (3),
-    QUORUM      (4),
-    ALL         (5),
-    LOCAL_QUORUM(6),
-    EACH_QUORUM (7),
-    SERIAL      (8),
-    LOCAL_SERIAL(9);
+    ANY          (0),
+    ONE          (1),
+    TWO          (2),
+    THREE        (3),
+    QUORUM       (4),
+    ALL          (5),
+    LOCAL_QUORUM (6),
+    EACH_QUORUM  (7),
+    SERIAL       (8),
+    LOCAL_SERIAL (9),
+    LOCAL_ONE   (10);
 
     // Used by the native protocol
     final int code;