+++ driver-core/CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 ------
 
 - [new] OSGi bundle (JAVA-142)
+- [improvement] Make collections returned by Row immutable (JAVA-205)
 - [bug] Don't retain unused PreparedStatement in memory (JAVA-201)
 - [bug] Add missing clustering order info in TableMetadata