+++ driver-core/CHANGELOG.rst
@@ -10,6 +10,8 @@ CHANGELOG
 Merged from 1.0 branch:
 
 - [new] OSGi bundle (JAVA-142)
+- [improvement] Make collections returned by Row immutable (JAVA-205)
+- [improvement] Limit internal thread pool size (JAVA-203)
 - [bug] Don't retain unused PreparedStatement in memory (JAVA-201)
 - [bug] Add missing clustering order info in TableMetadata
 - [bug] Allow bind markers for collections in the query builder (JAVA-196)
@@ -60,6 +62,7 @@ Merged from 1.0 branch:
 
 - [new] OSGi bundle (JAVA-142)
 - [improvement] Make collections returned by Row immutable (JAVA-205)
+- [improvement] Limit internal thread pool size (JAVA-203)
 - [bug] Don't retain unused PreparedStatement in memory (JAVA-201)
 - [bug] Add missing clustering order info in TableMetadata
 - [bug] Allow bind markers for collections in the query builder (JAVA-196)